Add files via upload
Stock market prediction is the act of trying to determine the future value of a company stock or other financial instrument traded on an exchange.

In this article, we will work with historical data about the stock prices of a publicly listed company. We will implement a mix of machine learning algorithms to predict the future stock price of this company, starting with simple algorithms like averaging and linear regression, and then move on to advanced techniques like Auto ARIMA and LSTM.

Table of Contents
Understanding the Problem Statement
Moving Average
Linear Regression
k-Nearest Neighbors
Auto ARIMA
Prophet
Long Short Term Memory (LSTM)

As you might have guessed, our focus will be on the technical analysis part. We’ll be using a dataset from Yahoo finance  (you can find historical data for various stocks here) and for this particular project, I have used the data for ‘Glenmark Pharmaceuticals Ltd’.